can now read JEGAGlobal.log
OptimizationResults class can now read either the log file piped from DAKOTA stdout OR THE JEGAGlobal log file. I had to handle an edge case where DAKOTA gets restarted mid-optimization, in which case the generations that were completed the first ime are repeated. Fortunately, I happened to be testing on a file where this was the case and the check for this will not break the functionality for using DAKOTA stdout file instead